Add validacion productos solo admin puede verlos Add comentarios Add new logo Add new url Update DB Add nuevo campo json Cambio nombre app

+++ app/src/main/java/com/example/inventariosapp/domain/model/login/LoginResponseModel.kt
@@ -16,5 +16,5 @@ data class LoginResponseModel(
     @SerializedName("usuarioSesionId") var usuarioSesionId : Int? = null,
     @SerializedName("fechaIngreso") var fechaIngreso : String? = null,
     @SerializedName("fechaModifico") var fechaModifico : String? = null,
-    @SerializedName("porcentajeComision") var porcentajeComision : Int? = null
+    @SerializedName("porcentajeComision") var porcentajeComision : Double? = null
 )
